Add preferred_client_key kdcpreauth callback
Add a new kdcpreauth callback which gets the preferred client key,
matching the etype-info2 entry we send to the client.  This callback
can be used by preauth mechs which need to use the client key in a
challenge sent by the KDC, now that we send only one etype-info2
entry.

ticket: 8199