Update for draft changes
Include the group number in the PRF+ input string for the derivation
of w.

Include the group number and enctype numbers in the PRF+ input string
for the derivation of K'[n].

Use "SPAKEkey" in the K'[n] pepper string instead of "SPAKEKey" for
case consistency with "SPAKEsecret" in the derivation of w.